Arreglado guardado de resultados en la bd
