+++ frontend/src/lib/api.ts
@@ -445,37 +445,28 @@ export interface GameSessionResponse {
 
 
 /**
-* Tipos de juegos disponibles
+ * Datos de una ronda del juego 1: Elegir Número
  */
-export type GameKey = 'touch_number' | 'order_sequence' | 'distribute_equal' | 'remove_equal';
-
-
-/**
- * Datos de una ronda de juego base
- */
-export interface RoundResultBase {
+export interface RoundResultGame1 {
   round: number;
   numbers: number[];
+  selected_number: number;
+  correct_number: number | null;
   is_correct: boolean;
   time_seconds: number;
 }
 
 
 /**
- * Datos de una ronda del juego 1: Elegir Número
- */
-export interface TouchNumber extends RoundResultBase {
-  selected_number: number;
-  correct_number: number | null;
-}
-
-
-/**
  * Datos de una ronda del juego 2: Ordenar Secuencia
  */
-export interface OrderSequence extends RoundResultBase {
+export interface RoundResultGame2 {
+  round: number;
+  numbers: number[];
   user_order: number[];
   correct_order: number[];
+  is_correct: boolean;
+  time_seconds: number;
 }
 
 
@@ -534,6 +525,35 @@ export const gamesAPI = {
   },
 
   /**
+ * Guarda el resultado de una ronda individual dentro de una sesión.
+ *
+ * Flujo de ejecución:
+ * 1. Envía los datos de la ronda (números, respuesta, corrección, tiempo)
+ * 2. El backend actualiza el campo results.attempts[] en game_sessions
+ * 3. Incrementa contadores de total_correct o total_incorrect según resultado
+ *
+ * @param sessionId - ID de la sesión activa donde guardar
+ * @param roundResult - Objeto con todos los datos de la ronda
+ * @returns Promesa que resuelve cuando se guarda exitosamente
+ *
+ * @example
+ * await gamesAPI.saveRoundResult('session-123', {
+ *   round: 1,
+ *   numbers: [1, 3, 5, 7, 9],
+ *   selected_number: 5,
+ *   correct_number: 5,
+ *   is_correct: true,
+ *   time_seconds: 12.5
+ * });
+ */
+  saveRoundResultGame1: async (sessionId: string, roundResult: RoundResultGame1): Promise<void> => {
+    await api.post(`/api/games/sessions/${sessionId}/round`, {
+      round_result: roundResult
+    });
+  },
+
+
+  /**
    * Guarda el resultado de una ronda individual dentro de una sesión.
    *
    * Flujo de ejecución:
@@ -555,7 +575,7 @@ export const gamesAPI = {
    *   time_seconds: 12.5
    * });
    */
-  saveRoundResult: async (sessionId: string, roundResult: TouchNumber | OrderSequence): Promise<void> => {
+  saveRoundResultGame2: async (sessionId: string, roundResult: RoundResultGame2): Promise<void> => {
     await api.post(`/api/games/sessions/${sessionId}/round`, {
       round_result: roundResult
     });
