+++ frontend/src/pages/games/Game1/Game1.css
@@ -23,53 +23,13 @@
 /* ========================================
    Números disponibles
    ======================================== */
-/* .available-numbers {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1.5rem;
-    justify-content: center;
-    padding: 2rem;
-    background: white;
-    border-radius: 16px;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-    min-height: 120px;
-}
-
-.game1-droppable-slot {
-    width: 90px;
-    height: 90px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: all 0.2s;
-}
-
-.game1-droppable-slot-hover {
-    transform: scale(1.05);
-}
-
-/* Espacio vacío (círculo gris) */
-/* .game1-empty-slot {
-    width: 90px;
-    height: 90px;
-    border-radius: 50%;
-    background: #d1d5db;
-    border: 3px dashed #9ca3af;
-    transition: all 0.2s;
-}
-
-.game1-droppable-slot-hover .game1-empty-slot {
-    background: #a5a5a5;
-    border-color: #a5a5a5;
-    border-width: 4px;
-    transform: scale(1.1);
-} */
 
 .numbers-grid {
     width: 100%;
     max-width: 1500px;
     margin: 20px auto;
     padding: 0px;
+    background: white;
 }
 
 .numbers-grid ion-row {
@@ -79,97 +39,6 @@
     gap: 60px;
     padding: 0;
 }
-
-/* ========================================
-   Tarjetas de números
-   ======================================== */
-/* .number-card-v2 {
-    width: 90px;
-    height: 90px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: #6a98e7;
-    color: white;
-    font-size: 2.5rem;
-    font-weight: bold;
-    border-radius: 50%;
-    box-shadow: 0 4px 12px rgba(59, 130, 246, 0.3);
-    cursor: grab;
-    transition: transform 0.2s, box-shadow 0.2s;
-    user-select: none;
-    position: relative;
-    overflow: hidden;
-}
-
-/* Estilo para tarjetas con pictograma */
-/* .number-card-pictogram {
-    background: #6a98e7;
-    border: 3px solid #6a98e7;
-}
-
-.number-card-pictogram.number-card-available {
-    border-color: #6a98e7;
-} */
-
-/* Imagen del pictograma */
-/* .pictogram-image {
-    width: 75px;
-    height: 75px;
-    object-fit: contain;
-    pointer-events: none;
-} */
-
-/* Números disponibles arriba (color gris) */
-/* .number-card-available {
-    background: #84acf1;
-    box-shadow: 0 4px 12px rgba(156, 163, 175, 0.3);
-} */
-
-/* Números bloqueados (ayuda pre-colocada) */
-/* .number-card-locked {
-    background: #10b981 !important;
-    cursor: not-allowed !important;
-    opacity: 0.9;
-    border: 3px solid #059669 !important;
-} */
-
-/* .number-card-locked.number-card-pictogram {
-    border: 3px solid #10b981 !important;
-}
-
-.number-card-v2:hover {
-    transform: translateY(-4px) scale(1.05);
-    box-shadow: 0 6px 16px rgba(59, 130, 246, 0.4);
-}
-
-.number-card-available:hover {
-    box-shadow: 0 6px 16px rgba(156, 163, 175, 0.4);
-}
-
-.number-card-locked:hover {
-    transform: none !important;
-    box-shadow: 0 4px 12px rgba(16, 185, 129, 0.4) !important;
-}
-
-.number-card-v2:active {
-    cursor: grabbing;
-}
-
-.number-card-locked:active {
-    cursor: not-allowed !important;
-}
-
-.number-card-dragging {
-    opacity: 0.5;
-    transform: scale(1.1);
-    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.4);
-}
-
-.number-value {
-    font-family: 'Arial', sans-serif;
-    z-index: 1;
-}  */
 
 .number-circle {
     display: flex;
